(#210) trocando integer para biginteger

+++ app/model/professor.py
@@ -6,7 +6,7 @@ class Professor(db.Model):
     __tablename__ = "professor"
 
     id_professor = db.Column(db.Integer, primary_key=True, autoincrement=True)
-    reg_professor = db.Column(db.Integer, nullable=True, unique=True)
+    reg_professor = db.Column(db.BigInteger, nullable=True, unique=True)
 
     name = db.Column(db.String(255), nullable=False)
     email = db.Column(db.String(255), nullable=True, unique=True)
